project_django/backend/requirements.txt
djangorestframework
markdown
django-filter
django-cors-headers
psycopg[binary,pool]
faker
gunicorn
